+++ extensions/CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- `checkPackages` function to check a list of packages against the default policy
+
 ## 5.0.0 - 2023-04-13
 
 ### Added